Fully implemented and tested
A correct implementation of a red black tree that generates an image of the tree. Supports inserting a value, searching for a value, printing all values in the tree in sorted order, and finding the overall min and max of the tree. Delete performs the binary search tree delete operation, but the red-black delete fixes have not been implemented yet.